STATInfo

+++ Memento Mori/Assets/Script/UnitScript/UnitType/Celte/GuerrierNu.cs
@@ -8,7 +8,7 @@ public class GuerrierNu :  IUnitType
     public int moral{get;set;}
     public int armure{get;set;}
     public int degat{get;set;}
-    public int vitesse{get;set;}
+    public float vitesse{get;set;}
     public GuerrierNu ()
     {
             vie = 160;
