+++ Memento Mori/Assets/Script/UnitScript/UnitType/IUnitType.cs
@@ -10,6 +10,6 @@ public interface IUnitType
     public int moral{get;set;}
     public int armure{get;set;}
     public int degat{get;set;}
-    public int vitesse{get;set;}
+    public float vitesse{get;set;}
 
 }
